+++ modules/bitgo/tsconfig.json
@@ -3,11 +3,19 @@
   "compilerOptions": {
     "outDir": "./dist",
     "rootDir": "./",
-    "typeRoots": ["./types", "../../types", "./node_modules/@types", "../../node_modules/@types"],
+    "typeRoots": [
+      "./types",
+      "../../types",
+      "./node_modules/@types",
+      "../../node_modules/@types"
+    ],
     "declarationDir": "./dist/types",
     "esModuleInterop": false
   },
-  "include": ["src/**/*", "package.json"],
+  "include": [
+    "src/**/*",
+    "package.json"
+  ],
   "typedocOptions": {
     "mode": "modules",
     "out": "docs",
@@ -26,6 +34,9 @@
     },
     {
       "path": "../sdk-api"
+    },
+    {
+      "path": "../sdk-coin-aca"
     },
     {
       "path": "../sdk-coin-ada"
