+++ modules/sdk-coin-aca/.eslintignore
@@ -1,0 +1,5 @@
+node_modules
+.idea
+public
+dist
+
