+++ modules/sdk-coin-aca/.prettierignore
@@ -1,0 +1,2 @@
+.nyc_output/
+dist/
