+++ modules/sdk-coin-aca/.prettierrc.yml
@@ -1,0 +1,3 @@
+printWidth: 120
+singleQuote: true
+trailingComma: 'es5'
