+++ modules/sdk-coin-aca/src/lib/address.ts
@@ -1,0 +1,5 @@
+import { BaseAddress } from '@bitgo/sdk-core';
+
+export class Address implements BaseAddress {
+  address: string;
+}
