+++ modules/sdk-coin-aca/src/lib/enum.ts
@@ -1,0 +1,3 @@
+export enum ContractType {}
+
+export enum PermissionType {}
