+++ modules/sdk-coin-aca/src/lib/index.ts
@@ -1,0 +1,13 @@
+// import * as Utils from './utils';
+// import * as Interface from './iface';
+
+// import dot keypair and export that
+// import dot transaction and export that
+
+export { Address } from './address';
+export { KeyPair } from './keyPair';
+export { Transaction } from './transaction';
+export { TransactionBuilder } from './transactionBuilder';
+export { TransactionBuilderFactory } from './transactionBuilderFactory';
+
+// export { Utils, Interface };
