+++ modules/sdk-coin-aca/src/lib/keyPair.ts
@@ -1,0 +1,2 @@
+// WE DONT NEED THIS?
+// WE CAN EXPORT THE POLKADOT KEYPAIR CLASS SINCE ITS THE SAME?
