+++ modules/sdk-coin-aca/src/lib/transactionBuilder.ts
@@ -1,0 +1,1 @@
+// I DONT THINK WE NEED THIS. WE CAN USE THE ABSTRACT TRANSACTIONBUILDER in DOT account-lib. IMPORT IT HERE?
