+++ modules/sdk-coin-aca/src/taca.ts
@@ -1,0 +1,18 @@
+// /**
+//  * Testnet Aca
+//  *
+//  * @format
+//  */
+// import { BaseCoin, BitGoBase } from '@bitgo/sdk-core';
+// import { BaseCoin as StaticsBaseCoin } from '@bitgo/statics';
+// import { Aca } from './aca';
+//
+// export class Taca extends Aca {
+//   protected constructor(bitgo: BitGoBase, staticsCoin?: Readonly<StaticsBaseCoin>) {
+//     super(bitgo, staticsCoin);
+//   }
+//
+//   static createInstance(bitgo: BitGoBase, staticsCoin?: Readonly<StaticsBaseCoin>): BaseCoin {
+//     return new Taca(bitgo, staticsCoin);
+//   }
+// }
